backend/requirements.txt
fastapi==0.123.7
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
python-dotenv==1.2.1
pydantic==2.12.5
selenium==4.15.2
